Player Movement Prototype
Worked on Static and Physics Entity class. Implemented "physics". Adjusted the play scene to have nothing but a plane and a player who can move around.

Made a shape helper class to help speed up making geometry in the level that does not need to be imported in and can be made with three.js

NEEDS REWORK: The player is clamed to not pass by y=0 for now. So the plane is purely visual as of right now.

Advanced movement not done. Only simple walk and jump.